+++ doc/source/changelog.rst
@@ -7,6 +7,7 @@ Version 1.4
 New features
 ------------
 
+* New customer portal
 * Database size monitoring - new panel (in admin menu)
 * Sofia Gateway : add new chooce lists for selecting codecs
 * Add sip profile DTMF options (pass-rfc2833)